Create removeGrid & createNewgrid functions, & button event listener
the removeGrid function will remove all of the child nodes from its
parent container to make room for the new grid being chosen. The
create new grid will accept the number the user chooses and greate a new
grid with with that number. Once the button is clicked the promt will
show for the users choice, then the removeGrid and createNewgrid
functions will call.